pyproject.toml
[tool.poetry]
name = "judgeflow"
version = "0.1.0"
description = "Evaluation framework for LLM outputs"
authors = ["keithtyser <keithtyser@gmail.com>"]
readme = "README.md"
packages = [{include = "judgeflow", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
openai = "^1.30.1"
pydantic = "^2.7.1"
sqlmodel = "^0.0.18"
typer = {extras = ["all"], version = "^0.9.0"}
pandas = "*"
PyYAML = "^6.0"
tabulate = "^0.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
judgeflow = "judgeflow.cli:app" 